added messages link to main and navbar, renamed directory events to event
