+++ app/javascript/react/components/event/EventIndexContainer.js
