+++ app/javascript/react/components/message/MessageIndexContainer.js
@@ -1,0 +1,14 @@
+import React from "react"
+
+const MessageIndexContainer = (props) => {
+
+  return (
+    <div>
+      <h2> Your messages are here: </h2>
+
+    </div>
+  )
+
+}
+
+export default MessageIndexContainer
